@@ -1,6 +1,6 @@
 {
-  "name": "@ro/jest-styled-components",
-  "version": "8.0.0",
+  "name": "jest-styled-components",
+  "version": "7.2.0",
   "description": "Jest utilities for Styled Components",
   "main": "./src/index.js",
   "typings": "./typings/index.d.ts",
@@ -10,49 +10,51 @@
     "src",
     "typings"
   ],
-  "repository": "git@github.com:styled-components/jest-styled-components.git",
+  "repository": "github:styled-components/jest-styled-components",
+  "bugs": {
+    "url": "https://github.com/styled-components/jest-styled-components/issues"
+  },
   "author": "Michele Bertoli",
   "license": "MIT",
   "scripts": {
-    "test": "yarn test:web && yarn test:native && yarn test:preact && yarn test:ts",
-    "test:web": "jest",
+    "test": "yarn test:web && yarn test:native && yarn test:preact",
+    "test:web": "jest --color",
     "test:web:watch": "jest --watch",
-    "test:native": "jest --config .jest.native.json",
-    "test:preact": "jest --config .jest.preact.json",
-    "test:ts": "tsc --strict ./typings/index.d.ts"
+    "test:native": "jest --color --config .jest.native.json",
+    "test:preact": "jest --color --config .jest.preact.json"
   },
   "devDependencies": {
     "@babel/core": "^7.0.0",
-    "@babel/plugin-proposal-class-properties": "^7.0.0",
-    "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
-    "@babel/preset-env": "^7.0.0",
+    "@babel/eslint-parser": "^7.16.3",
+    "@babel/preset-env": "^7.16.4",
     "@babel/preset-react": "^7.0.0",
-    "@testing-library/react": "^9.3.2",
-    "babel-core": "^7.0.0-bridge.0",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^24.8.0",
+    "@testing-library/react": "^12.1.2",
+    "@types/styled-components": "^5.1.15",
+    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.5",
+    "babel-jest": "^27.3.1",
+    "babel-plugin-styled-components": "^2.0.1",
     "enzyme": "^3.6.0",
-    "enzyme-adapter-preact-pure": "^2.2.0",
-    "enzyme-adapter-react-16": "^1.5.0",
+    "enzyme-adapter-preact-pure": "^3.3.0",
     "enzyme-to-json": "^3.3.1",
-    "eslint": "^6.6.0",
-    "eslint-config-prettier": "^6.5.0",
-    "husky": "^3.0.9",
-    "jest": "^24.8.0",
-    "lint-staged": "^9.4.2",
+    "eslint": "^8.3.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "husky-v4": "^4.3.8",
+    "jest": "^27.3.1",
+    "lint-staged": "^12.1.1",
     "preact": "^10.0.0-beta.2",
     "preact-render-to-json": "^3.6.6",
-    "prettier": "^1.18.2",
-    "react": "^16.8.0",
-    "react-dom": "^16.8.0",
-    "react-is": "^16.8.6",
-    "react-native": "^0.61.4",
-    "react-test-renderer": "^16.8.0",
-    "styled-components": "^6.0.0",
-    "typescript": "^3.0.3"
+    "prettier": "^2.3.2",
+    "pretty-format": "^27.5.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-is": "^17.0.2",
+    "react-native": "^0.66.3",
+    "react-test-renderer": "^17.0.2",
+    "styled-components": "^6.0.9"
   },
   "dependencies": {
-    "css": "^2.2.4"
+    "@adobe/css-tools": "^4.0.1"
   },
   "peerDependencies": {
     "styled-components": ">= 5"
@@ -64,8 +66,7 @@
   },
   "lint-staged": {
     "*.js": [
-      "eslint --fix",
-      "git add"
+      "eslint --fix"
     ]
   },
   "jest": {
@@ -78,9 +79,13 @@
     "snapshotSerializers": [
       "enzyme-to-json/serializer"
     ],
+    "testEnvironment": "jsdom",
     "testPathIgnorePatterns": [
       "<rootDir>/test/native",
       "<rootDir>/test/preact"
     ]
+  },
+  "engines": {
+    "node": ">= 12"
   }
 }
