@@ -1,68 +1,63 @@
-const { getCSS, matcherTest, buildReturnMessage } = require("./utils");
+const { getCSS, matcherTest, buildReturnMessage } = require('./utils');
 
-const shouldDive = node =>
-  typeof node.dive === "function" && typeof node.type() !== "string";
+const shouldDive = (node) => typeof node.dive === 'function' && typeof node.type() !== 'string';
 
-const isTagWithClassName = node =>
-  node.exists() && node.prop("className") && typeof node.type() === "string";
+const isTagWithClassName = (node) => node.exists() && node.prop('className') && typeof node.type() === 'string';
 
-const isStyledClass = className =>
-  /^(\w+(-|_))?sc-/.test(className);
+const isStyledClass = (className) => /(_|-)+sc-.+|^sc-/.test(className);
 
-const hasClassName = node =>
-  node.length > 0
-  && typeof node.props === "function"
-  && node.prop("className")
-  && isStyledClass(node.prop("className"))
+const hasClassName = (node) =>
+  node.length > 0 &&
+  typeof node.props === 'function' &&
+  node.prop('className') &&
+  isStyledClass(node.prop('className'));
 
-const getClassNames = received => {
+const getClassNames = (received) => {
   let className;
 
   if (received) {
-    if (received.$$typeof === Symbol.for("react.test.json")) {
+    if (received.$$typeof === Symbol.for('react.test.json')) {
       className = received.props.className || received.props.class;
     } else if (hasClassName(received)) {
-      className = received.prop("className");
-    } else if (typeof received.exists === "function" && received.exists()) {
+      className = received.prop('className');
+    } else if (typeof received.exists === 'function' && received.exists()) {
       const tree = shouldDive(received) ? received.dive() : received;
       const components = tree.findWhere(isTagWithClassName);
       if (components.length) {
-        className = components.first().prop("className");
+        className = components.first().prop('className');
       }
     } else if (global.Element && received instanceof global.Element) {
-      className = Array.from(received.classList).join(" ");
+      className = Array.from(received.classList).join(' ');
     }
   }
 
   return className ? className.split(/\s/) : [];
 };
 
-const hasAtRule = options =>
-  Object.keys(options).some(option => ["media", "supports"].includes(option));
+const hasAtRule = (options) => Object.keys(options).some((option) => ['media', 'supports'].includes(option));
 
 const getAtRules = (ast, options) => {
-  const mediaRegex = /(\([a-z-]+:)\s?([a-z0-9.]+\))/g;
-
   return Object.keys(options)
-    .map(option =>
+    .map((option) =>
       ast.stylesheet.rules
-        .filter(
-          rule =>
-            rule.type === option &&
-            rule[option] === options[option].replace(mediaRegex, "$1$2")
-        )
-        .map(rule => rule.rules)
+        .filter((rule) => rule.type === option && rule[option] === options[option].replace(/:\s/g, ":"))
+        .map((rule) => rule.rules)
         .reduce((acc, rules) => acc.concat(rules), [])
     )
     .reduce((acc, rules) => acc.concat(rules), []);
 };
 
-const getModifiedClassName = (className, staticClassName, modifier = "") => {
+/** stylis v4 renders descendant selectors without a trailing space sometimes which trips up detection */
+const removeSpaceAfterSelector = input => input.replace(/([>~+]) +/g, '$1')
+
+const normalizeQuotations = (input) => input.replace(/['"]/g, '"');
+
+const getModifiedClassName = (className, staticClassName, modifier = '') => {
   const classNameSelector = `.${className}`;
-  let prefix = "";
+  let prefix = '';
 
   modifier = modifier.trim();
-  if (modifier.includes("&")) {
+  if (modifier.includes('&')) {
     modifier = modifier
       // & combined with other selectors and not a precedence boost should be replaced with the static className, but the first instance should be the dynamic className
       .replace(/(&[^&]+?)&/g, `$1.${staticClassName}`)
@@ -71,68 +66,53 @@ const getModifiedClassName = (className, staticClassName, modifier = "") => {
     prefix += classNameSelector;
   }
   const first = modifier[0];
-  if (first !== ":" && first !== "[") {
-    prefix += " ";
+  if (first !== ':' && first !== '[') {
+    prefix += ' ';
   }
 
   return `${prefix}${modifier}`.trim();
 };
 
 const hasClassNames = (classNames, selectors, options) => {
-  const staticClassNames = classNames.filter(x => isStyledClass(x));
+  const staticClassNames = classNames.filter((x) => isStyledClass(x));
 
-  return classNames.some(className =>
-    staticClassNames.some(staticClassName =>
-      selectors.includes(
-        getModifiedClassName(
-          className,
-          staticClassName,
-          options.modifier
-        ).replace(/['"]/g, '"')
+  return classNames.some((className) =>
+    staticClassNames.some((staticClassName) =>
+      selectors.map(removeSpaceAfterSelector).includes(
+        removeSpaceAfterSelector(normalizeQuotations(getModifiedClassName(className, staticClassName, options.modifier).replace(/['"]/g, '"')))
       )
     )
   );
 };
 
 const getRules = (ast, classNames, options) => {
-  const rules = hasAtRule(options)
-    ? getAtRules(ast, options)
-    : ast.stylesheet.rules;
+  const rules = (hasAtRule(options) ? getAtRules(ast, options) : ast.stylesheet.rules).map((rule) => ({
+    ...rule,
+    selectors: Array.isArray(rule.selectors) ? rule.selectors.map(normalizeQuotations) : rule.selectors,
+  }));
 
-  return rules.filter(
-    rule =>
-      rule.type === "rule" && hasClassNames(classNames, rule.selectors, options)
-  );
+  return rules.filter((rule) => rule.type === 'rule' && hasClassNames(classNames, rule.selectors, options));
 };
 
-const handleMissingRules = options => ({
+const handleMissingRules = (options) => ({
   pass: false,
   message: () =>
-    `No style rules found on passed Component${
-      Object.keys(options).length
-        ? ` using options:\n${JSON.stringify(options)}`
-        : ""
-    }`
+    `No style rules found on passed Component${Object.keys(options).length ? ` using options:\n${JSON.stringify(options)}` : ''
+    }`,
 });
 
 const getDeclaration = (rule, property) =>
   rule.declarations
-    .filter(
-      declaration =>
-        declaration.type === "declaration" && declaration.property === property
-    )
+    .filter((declaration) => declaration.type === 'declaration' && declaration.property === property)
     .pop();
 
-const getDeclarations = (rules, property) =>
-  rules.map(rule => getDeclaration(rule, property)).filter(Boolean);
+const getDeclarations = (rules, property) => rules.map((rule) => getDeclaration(rule, property)).filter(Boolean);
 
-const normalizeOptions = options =>
+const normalizeOptions = (options) =>
   options.modifier
     ? Object.assign({}, options, {
-        modifier: Array.isArray(options.modifier)
-          ? options.modifier.join("")
-          : options.modifier
-      })
+      modifier: Array.isArray(options.modifier) ? options.modifier.join('') : options.modifier,
+    })
     : options;
 
 function toHaveStyleRule(component, property, expected, options = {}) {
@@ -148,14 +128,11 @@ function toHaveStyleRule(component, property, expected, options = {}) {
   const declarations = getDeclarations(rules, property);
   const declaration = declarations.pop() || {};
   const received = declaration.value;
-  const pass =
-    !received && !expected && this.isNot
-      ? false
-      : matcherTest(received, expected);
+  const pass = matcherTest(received, expected, this.isNot);
 
   return {
     pass,
-    message: buildReturnMessage(this.utils, pass, property, received, expected)
+    message: buildReturnMessage(this.utils, pass, property, received, expected),
   };
 }
 
