@@ -80,6 +80,18 @@ it('negated ".not" modifier with value', () => {
   expect(renderer.create(<Button />).toJSON()).not.toHaveStyleRule('padding', undefined);
 });
 
+it('negated ".not" modifier fails when rule present with no value being asserted', () => {
+  const Button = styled.Text`
+    background-color: blue;
+  `;
+
+  expect(renderer.create(<Button />).toJSON()).toHaveStyleRule('background-color', 'blue');
+  expect(renderer.create(<Button />).toJSON()).not.toHaveStyleRule('opacity');
+  expect(() => {
+    expect(renderer.create(<Button />).toJSON()).not.toHaveStyleRule('background-color');
+  }).toThrowError();
+});
+
 it('jest asymmetric matchers', () => {
   const Button = styled.Text`
     background-color: ${({ transparent }) => (transparent ? 'transparent' : 'papayawhip')};
@@ -114,8 +126,8 @@ it('theming', () => {
     padding: 0.25px 1px;
     border-radius: 3px;
 
-    color: ${props => props.theme.main};
-    border: 2px solid ${props => props.theme.main};
+    color: ${(props) => props.theme.main};
+    border: 2px solid ${(props) => props.theme.main};
   `;
 
   Button.defaultProps = {
@@ -137,4 +149,14 @@ it('theming', () => {
   );
 
   expect(renderer.create(component).toJSON()).toHaveStyleRule('color', 'mediumseagreen');
+});
+
+it('style prop is an object', () => {
+  const StyledView = styled.TouchableOpacity`
+    background-color: papayawhip;
+  `;
+  const tree = renderer.create(<StyledView />).toJSON();
+
+  expect(typeof tree.props.style).toBe('object')
+  expect(tree).toHaveStyleRule('background-color', 'papayawhip');
 });
