@@ -3,8 +3,11 @@ import { mount, shallow } from 'enzyme';
 import React from 'react';
 import renderer from 'react-test-renderer';
 import styled, { ThemeContext, ThemeProvider } from 'styled-components';
+import prettyFormat, { plugins } from 'pretty-format';
+import {setStyleSheetSerializerOptions} from '../serializer';
+import styleSheetSerializer from '../src/styleSheetSerializer';
 
-const toMatchSnapshot = component => {
+const toMatchSnapshot = (component) => {
   expect(renderer.create(component).toJSON()).toMatchSnapshot('react-test-renderer');
   expect(shallow(component)).toMatchSnapshot('shallow');
   expect(mount(component)).toMatchSnapshot('mount');
@@ -17,6 +20,31 @@ const shallowWithTheme = (tree, theme) => {
 
   return shallow(tree);
 };
+
+/**
+ * Serializes a value similarly to jest-snapshot
+ *
+ * Imitates the serialization logic inside jest-snapshot,
+ * so we can pass options, including indent, to it.
+ *
+ * @see https://github.com/facebook/jest/blob/615084195ae1ae61ddd56162c62bbdda17587569/packages/jest-snapshot/src/utils.ts#L155-L163
+ * @see https://github.com/facebook/jest/blob/615084195ae1ae61ddd56162c62bbdda17587569/packages/jest-snapshot/src/plugins.ts#L24-L32
+ */
+const serialize = (val, indent = 2) =>
+  prettyFormat(val, {
+    escapeRegex: true,
+    indent,
+    plugins: [
+      styleSheetSerializer,
+      plugins.ReactTestComponent,
+      plugins.ReactElement,
+      plugins.DOMElement,
+      plugins.DOMCollection,
+      plugins.Immutable,
+      plugins.AsymmetricMatcher,
+    ],
+    printFunctionName: false,
+  });
 
 it('null', () => {
   expect(null).toMatchSnapshot();
@@ -126,8 +154,8 @@ it('theming', () => {
     padding: 0.25em 1em;
     border-radius: 3px;
 
-    color: ${props => props.theme.main};
-    border: 2px solid ${props => props.theme.main};
+    color: ${(props) => props.theme.main};
+    border: 2px solid ${(props) => props.theme.main};
   `;
 
   Button.defaultProps = {
@@ -152,7 +180,7 @@ it('theming', () => {
 
 it('shallow with theme', () => {
   const Button = styled.button`
-    color: ${props => props.theme.main};
+    color: ${(props) => props.theme.main};
   `;
 
   const theme = {
@@ -238,12 +266,44 @@ it('referring to other components', () => {
   `;
 
   const component = (
-    <Link href="#">
-      <Icon />
-      <Label>Hovering my parent changes my style!</Label>
-      <TextWithConditionalFormatting>I should be green</TextWithConditionalFormatting>
-      <TextWithConditionalFormatting error>I should be red</TextWithConditionalFormatting>
-    </Link>
+    <Container>
+      <Link href="#">
+        <Icon />
+        <Label>Hovering my parent changes my style!</Label>
+        <TextWithConditionalFormatting>I should be green</TextWithConditionalFormatting>
+        <TextWithConditionalFormatting error>I should be red</TextWithConditionalFormatting>
+      </Link>
+    </Container>
+  );
+
+  expect(renderer.create(component).toJSON()).toMatchSnapshot('react-test-renderer');
+  expect(mount(component)).toMatchSnapshot('mount');
+  expect(render(component).container.firstChild).toMatchSnapshot('react-testing-library');
+});
+
+it('strips unused styles', () => {
+  const BlueText = styled.div`
+    color: blue;
+  `;
+  const RedText = styled.div`
+    color: red;
+  `;
+
+  const FancyText = styled.div`
+    font-size: 16px;
+
+    ${BlueText} {
+      text-align: right;
+    }
+    ${RedText} {
+      text-align: center;
+    }
+  `;
+
+  const component = (
+    <FancyText>
+      <BlueText>A container that has styles for an unused child</BlueText>
+    </FancyText>
   );
 
   expect(renderer.create(component).toJSON()).toMatchSnapshot('react-test-renderer');
@@ -254,16 +314,94 @@ it('referring to other components', () => {
 it('referring to other unreferenced components', () => {
   const UnreferencedLink = styled.a`
     font-size: 1.5em;
-  `
+  `;
 
   const ReferencedLink = styled(UnreferencedLink)`
     color: palevioletred;
     font-weight: bold;
-  `
+  `;
 
   toMatchSnapshot(
     <div>
       <ReferencedLink>Styled, exciting Link</ReferencedLink>
     </div>
   );
+});
+
+it('should match the same snapshot rerendering the same element', () => {
+  const StyledDiv = styled.div`
+    color: palevioletred;
+    font-weight: bold;
+  `;
+
+  const {rerender, container} = render(<StyledDiv />);
+
+  expect(container).toMatchSnapshot('first-render');
+  rerender(<StyledDiv />);
+  expect(container).toMatchSnapshot('second-render');
+});
+
+it('allows to disable css snapshotting', () => {
+  setStyleSheetSerializerOptions({ addStyles: false })
+  const A = styled.div`
+    color: red;
+  `; const B = styled.div`
+    color: red;
+  `;
+
+  toMatchSnapshot(
+    <div>
+      <A>Styled, exciting div</A>
+      <B>Styled, exciting div</B>
+    </div>
+  );
+});
+
+it('allows to set a css classNameFormatter', () => {
+  setStyleSheetSerializerOptions({ classNameFormatter: (index) => `styledComponent${index}`  })
+  const A = styled.div`
+    color: red;
+  `;
+  const B = styled.div`
+    color: green;
+  `;
+
+  toMatchSnapshot(
+    <div>
+      <A>Styled, exciting div</A>
+      <B>Styled, exciting div</B>
+    </div>
+  );
+});
+
+it('responds to indent configuration', () => {
+  const Link = styled.a`
+    color: blue;
+  `;
+
+  const mounted = renderer.create(<Link>Styled, exciting Link</Link>);
+
+  expect(serialize(mounted)).toMatchInlineSnapshot(`
+    ".styledComponent0 {
+      color: blue;
+    }
+
+    <a
+      className=\\"styledComponent0\\"
+    >
+      Styled, exciting Link
+    </a>"
+  `);
+
+  expect(serialize(mounted, 0)).toMatchInlineSnapshot(`
+    ".styledComponent0 {
+    color: blue;
+    }
+
+    <a
+    className=\\"styledComponent0\\"
+    >
+    Styled, exciting Link
+    </a>"
+  `);
 });
